Formalize LLM API modeling, add research capabilities
