+++ .gitignore
@@ -15,3 +15,4 @@ conversations
 
 # Testing
 .tox
+personal
