--- experiences2.json
@@ -1,1 +1,0 @@
-{"contact": {"name": "Ansh Chaurasia", "phone": "+14086072856", "email": "eye.am.ansh@gmail.com", "location": "San Francisco, California", "social_networks": [{"network": "GitHub", "username": "AnshKetchum"}, {"network": "LinkedIn", "username": "AnshKetchum"}]}, "project_experiences": [{"name": "AI Art Generation", "location": "", "date": "Fall 2022", "description": "Fine-tuned Stability AI\u2019s Stable Diffusion XL text-to-image model using the Low Rank Adaptation (LoRA) technique. Used it to create concept anime art that was subsequently sold as an NFT and vector logos for branding projects.", "highlights": ["Fine-tuned Stability AI\u2019s Stable Diffusion XL text-to-image model using the Low Rank Adaptation (LoRA) technique. Used it to create concept anime art that was subsequently sold as an NFT and vector logos for branding projects."]}, {"name": "Voice Controlled AI Robot", "location": "", "date": "Spring 2023", "description": "Built a robot using embedded systems like Arduino. Leveraged data science methods like PCA for dimensionality reduction and voice matching to enable it to respond to voice commands.", "highlights": ["Built a robot using embedded systems like Arduino. Leveraged data science methods like PCA for dimensionality reduction and voice matching to enable it to respond to voice commands."]}], "work_experiences": [{"experience_start": "5-2022", "experience_end": "12-2023", "position": "Founding Engineer", "company": "Flockfysh Inc", "location": "", "description": "Built the founding software and concept for an AI system. Led the technical side of the pitch and secured the company a pre-seed round at a $1B valuation. Published an open-source package viewable on PyPI (refyre) to streamline local file management. Built a command-line interface allowing customers to access MongoDB and DigitalOcean data.", "highlights": ["Built the founding software and concept for an AI system. Led the technical side of the pitch and secured the company a pre-seed round at a 1B valuation. Published an open-source package viewable on PyPI (refyre) to streamline local file management. Built a command-line interface allowing customers to access MongoDB and DigitalOcean data."]}, {"experience_start": "5-2022", "experience_end": "12-2023", "position": "AI Software Engineer", "company": "Gamelore Inc", "location": "", "description": "Utilized TensorFlow 2.0 to contribute to an AI-based product that automates sports (tennis), dance, yoga, and physiotherapy coaching. Developed a YOLO model and trained it to perform action recognition with 85% accuracy. Developed ball detection algorithms in C++ and patented the technology for sports like tennis.", "highlights": ["Utilized TensorFlow 2.0 to contribute to an AI-based product that automates sports (tennis), dance, yoga, and physiotherapy coaching. Developed a YOLO model and trained it to perform action recognition with 85 percent accuracy. Developed ball detection algorithms in C++ and patented the technology for sports like tennis."]}, {"experience_start": "5-2022", "experience_end": "12-2023", "position": "AI Software Engineer", "company": "Ambrego Inc", "location": "", "description": "Worked on software development using Python to automate the process of calculating guide RNAs for various proteins. Utilized contemporary biotechnology AI software such as AlphaFold, Equidock, and Equibind for accuracy and prediction.", "highlights": ["Worked on software development using Python to automate the process of calculating guide RNAs for various proteins. Utilized contemporary biotechnology AI software such as AlphaFold, Equidock, and Equibind for accuracy and prediction."]}], "education": [{"institution": "UC Berkeley", "type": "degree", "start_date": "2014", "area": "EECS", "end_date": "2015", "relevant_coursework": ["Computer architecture", "Computer Security", "Integrated Circuit / Chip Design", "Machine Learning", "Mathematical Optimization Models", "Discrete Math and Probability Theory", "Structure and Interpretation of Computer Programs", "Designing Information Devices and Systems I", "Algorithms & Complexity"], "show_for_type": "all", "gpa": "4.5 UW", "study_type": "BS"}], "relevant_extracurriculars": [{"name": "", "description": "", "achievements": "", "keywords": ""}], "certificates": [{"name": "deeplearning.ai TensorFlow Developer Certification", "location": "Remote", "date": "Fall 2019", "highlights": ["Developed and achieved proficiency in building models in the computer vision, natural language processing, and time series domain."]}], "skills": [{"name": "Programming Languages", "details": "Python"}, {"name": "Programming Languages", "details": "C++"}, {"name": "Programming Languages", "details": "Java"}, {"name": "Frameworks", "details": "TensorFlow"}, {"name": "Frameworks", "details": "PyTorch"}, {"name": "Frameworks", "details": "Keras"}, {"name": "Tools", "details": "Git"}, {"name": "Tools", "details": "Docker"}, {"name": "Tools", "details": "Jenkins"}, {"name": "Other", "details": "Machine Learning"}, {"name": "Other", "details": "Computer Vision"}, {"name": "Other", "details": "Natural Language Processing"}, {"name": "Other", "details": "Deep Learning"}, {"name": "Other", "details": "Reinforcement Learning"}], "summary": "ExperiencedAISoftwareEngineerwithastrongbackgroundincomputerscienceandaproventrackrecordofbuildingandleadingtechnicalteams.Skilledinmachinelearning,computervision,andsoftwaredevelopment,withapassionforleveragingAItosolvereal-worldproblems.Provenabilitytoinnovateanddeliverresults,asdemonstratedbycontributionstofoundinganddevelopingAI-basedproducts,securingpre-seedfunding,andpublishingopen-sourcepackages."}
